POK-35 Renamed login service class to LoginEvent
