+++ PokeAppiOS/PokeAppiOS.csproj
@@ -68,7 +68,7 @@
         <Reference Include="System.Numerics.Vectors" />
     </ItemGroup>
     <ItemGroup>
-      <PackageReference Include="Xamarin.Essentials" Version="1.6.1" />
+      <PackageReference Include="Xamarin.Essentials" Version="1.7.5" />
       <PackageReference Include="Toast.iOS">
         <Version>1.2.0</Version>
       </PackageReference>
